+++ Pos/src/com/android/pos/report/cashflow/CashFlowBillDetailFragment.java
@@ -143,9 +143,9 @@ public class CashFlowBillDetailFragment extends BaseFragment {
 		mReferenceNoText.setText(mBill.getBillReferenceNo());
 		mDateText.setText(CommonUtil.formatDate(mBill.getBillDate()));
 
-		Integer totalAmount = mBill.getBillAmount();
-		Integer paymentAmount = mBill.getPayment();
-		Integer outstandingAmount = totalAmount - paymentAmount;
+		Float totalAmount = mBill.getBillAmount();
+		Float paymentAmount = mBill.getPayment();
+		Float outstandingAmount = totalAmount - paymentAmount;
 		
 		mSupplierText.setText(mBill.getSupplierName());
 		mDueDateText.setText(CommonUtil.formatDate(mBill.getBillDueDate()));
