+++ Pos/src/com/android/pos/report/outstanding/OutstandingBillListFragment.java
@@ -139,7 +139,7 @@ public class OutstandingBillListFragment extends BaseFragment
 		
 		setBackButtonVisible(false);
 		
-		Integer totalBills = 0;
+		Float totalBills = Float.valueOf(0);
 		
 		for (Bills bill : bills) {
 			totalBills += (bill.getBillAmount() - bill.getPayment());
